Retrieve only provisioned devices
It is possible for the Arlo session to return devices with a state of
'removed'. This occurs, for example, when a user removes a defective
device from the system for replacement. In this case, it is possible for
this call to return duplicate devices if the user named the newly
provisioned device with the same name. This would cause the user to
sometimes receive old data from a removed device instead of the newly
provisioned device.